Vendor DataSource Complete

+++ inventory/src/main/java/com/example/inventory/controller/HelloController.java
@@ -3,16 +3,15 @@ package com.example.inventory.controller;
 
 import com.example.inventory.data.DatabaseConnection;
 import com.example.inventory.data.GoodsDataSource;
+import com.example.inventory.data.VendorsDataSource;
 import com.example.inventory.repository.Goods;
+import com.example.inventory.repository.Vendor;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
-import javafx.scene.control.Button;
-import javafx.scene.control.TableColumn;
-import javafx.scene.control.TableView;
-import javafx.scene.control.TextField;
+import javafx.scene.control.*;
 import javafx.scene.control.cell.PropertyValueFactory;
 
 import java.net.URL;
@@ -23,11 +22,23 @@ import java.util.ResourceBundle;
 
 public class HelloController implements Initializable{
     @FXML
+    private TableView<Vendor> tvVendors;
+    @FXML
+    private TextField vtfname;
+    @FXML
+    private TextField vtfnumber;
+    @FXML
     private TextField tfname;
+    @FXML
+    private Button vsbtn;
+    @FXML
+    private Button vrbtn;
     @FXML
     private TextField tfcategory;
     @FXML
     private TextField tfquantity;
+    @FXML
+    private Label statuslb;
     @FXML
     private Button btnsave;
     @FXML
@@ -39,62 +50,77 @@ public class HelloController implements Initializable{
     @FXML
     private TableColumn<Goods, Integer> goodscolqty;
     @FXML
-    private TableColumn<Goods, String> vendorcolname;
+    private TableColumn<Vendor, String> vendorcolname;
     @FXML
-    private TableColumn<Goods, Integer> vendorcolnumber;
+    private TableColumn<Vendor, String > vendorcolnumber;
     @FXML
     private TableView<Goods> tvgoods;
     @FXML
-    private void saveHandleButtonAction(ActionEvent event) {
-        insertGoods();
-        getGoods();
-        showGoods();
+    private void handleButtonAction(ActionEvent event) {
+        if (event.getSource() == btnsave) {
+            insertGoods();
+            showGoods();
+            tfname.clear();
+            tfcategory.clear();
+            tfquantity.clear();
+        }
+        else if (event.getSource() == btnremove){
+            deleteGoods();
+            showGoods();
+            tfname.clear();
+            tfcategory.clear();
+            tfquantity.clear();
+        }
+
     }
     @FXML
-    private void removeHandleButtonAction(ActionEvent event){
-        deleteGoods();
-        getGoods();
-        showGoods();
+    private void onHandleVendorAction(ActionEvent event){
+        if (event.getSource() == vsbtn) {
+            insertVendors();
+            showVendors();
+            vtfname.clear();
+            vtfnumber.clear();
+
+        }
+        else if (event.getSource() == vrbtn){
+            deleteVendors();
+            showVendors();
+            vtfname.clear();
+            vtfnumber.clear();
+        }
     }
-
-
 
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
-        getGoods();
         showGoods();
+        showVendors();
     }
 
-    public void getGoods() {     // Retrieving goods from database
-        ObservableList<Goods> newGoods = FXCollections.observableArrayList();
 
-        DatabaseConnection connection = new DatabaseConnection();
-        Connection connectDB = connection.getConnection();
-
-        String connectQuery = "SELECT name, category, quantity FROM goods";
-
-
-        try {
-            Statement statement = connectDB.createStatement();
-            ResultSet queryOutput = statement.executeQuery(connectQuery);
-
-            while (queryOutput.next()) {
-                Goods goods = new Goods(queryOutput.getString("name"), queryOutput.getString("category"), queryOutput.getInt("quantity"));
-                newGoods.add(goods);
-            }
-
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        tvgoods.setItems(newGoods);
-    }
 
 
     public void showGoods(){
+        GoodsDataSource getGoodsList = new GoodsDataSource();
+        ObservableList<Goods> list = getGoodsList.getGoods();
+
+        tvgoods.setItems(list);
+
         goodscolcategory.setCellValueFactory(new PropertyValueFactory<>("category"));
         goodscolname.setCellValueFactory(new PropertyValueFactory<>("name"));
         goodscolqty.setCellValueFactory(new PropertyValueFactory<Goods, Integer>("quantity"));
     }
+
+    public void showVendors(){
+        VendorsDataSource getVendorList = new VendorsDataSource();
+        ObservableList<Vendor> list = getVendorList.getVendor();
+
+        tvVendors.setItems(list);
+
+        vendorcolname.setCellValueFactory(new PropertyValueFactory<>("name"));
+        vendorcolnumber.setCellValueFactory(new PropertyValueFactory<>("number"));
+
+    }
+
 
 
     public void insertGoods(){
@@ -102,6 +128,22 @@ public class HelloController implements Initializable{
         GoodsDataSource addToDatabase = new GoodsDataSource();
         addToDatabase.addGoods(tfname.getText(), tfcategory.getText(), Integer.parseInt(tfquantity.getText()) );
     }
+
+    public void insertVendors(){
+
+        VendorsDataSource addToDatabase = new VendorsDataSource();
+        addToDatabase.addVendors(vtfname.getText(), vtfnumber.getText() );
+    }
+
+
+
+
+    public void deleteVendors(){
+
+        VendorsDataSource removeFromDB = new VendorsDataSource();
+        removeFromDB.removeVendor(vtfname.getText(), vtfnumber.getText() );
+    }
+
 
     public void deleteGoods(){
         GoodsDataSource removeFromDB = new GoodsDataSource();
